feat : 정적 리소스 파일을 요청하는 컨트롤러에서 사용되는 반복적 연산에 대한 Util 클래스 구현
기존의 FilePathParser는 그 기능이 중복되므로 삭제
